refactor zaps
