+++ src/nostr/DefaultRelays.ts
@@ -1,9 +1,9 @@
 import { RelaySetting } from "./Types";
 
 export const defaultRelays: RelaySetting[]  = [
-    {relayUrl: "wss://nostr.wine", read: false, write: false},
-    {relayUrl: "wss://filter.nostr.wine", read: true, write: true},
-    {relayUrl: "wss://eden.nostr.land", read: true, write: false},
+    {relayUrl: "wss://nostr.wine", read: true, write: false},
+    {relayUrl: "wss://filter.nostr.wine", read: false, write: true},
+    {relayUrl: "wss://eden.nostr.land", read: false, write: false},
     {relayUrl: "wss://relay.nostr.band/", read: false, write: false},
     {relayUrl: "wss://nostr-pub.wellorder.net", read: false, write: true},
     {relayUrl: "wss://relay.nostr.band/all", read: false, write: true},
